requirements.txt
Flask==3.0.0
gspread==6.1.2
google-auth==2.34.0
qrcode==7.4.2
